requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
boto3==1.34.116
botocore==1.34.116
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.2
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
langcodes==3.4.0
language_data==1.2.0
llama-index==0.10.40
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.40
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.21
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-postgres==0.1.7
llama-parse==0.4.4
llamaindex-py-client==0.1.19
lxml==5.2.2
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.30.5
openpyxl==3.1.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pgvector==0.2.5
pillow==10.3.0
platformdirs==4.2.2
preshed==3.0.9
prompt_toolkit==3.0.46
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pypdf==4.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
s3transfer==0.10.1
setuptools==70.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
stack-data==0.6.3
striprtf==0.0.26
tenacity==8.3.0
thinc==8.2.4
tiktoken==0.7.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.9.4